Compatibility fix for different cifs implementations
There are several quirks with cifs.
Namely, some shares do add entity type mark (D for directories, A for files)
thus to avoid parsing everything that preceeds last 40 chars (leading to filenames like 'filename.bin      A')
a new validation for the existence of 'A' column was added, effectively concatenating every prior column into a single filename.
This is a compatibility fix and should not be regarded as a go-to solution.
The whole cifs implementation should be moved into an external cifs library.